requirements.txt
tensorflow==2.20.0
pandas
dvc
mlflow==3.2.0
notebook
numpy
matplotlib
seaborn
python-box==7.3.2
pyYAML
tqdm
ensure==1.0.4
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
tf-keras
fastapi
uvicorn
tensorflow
numpy
pillow
python-multipart

